move match.py into gatovid.api.game
